Bugs fixed before payment.

+++ ShoppingCart.py
@@ -5,6 +5,7 @@ from google.appengine.ext import ndb
 from math import sin, cos, sqrt, atan2, radians
 from UsersDB import UsersDB
 from ProductsDB import ProductsDB
+from VendorProductsDB import VendorProductsDB
 from PharmacyDB import PharmacyDB
 from CartDB import CartDB
 
@@ -33,27 +34,28 @@ class ShoppingCart(webapp2.RequestHandler):
                 self.redirect('/UserSignIn?notification=EmailIdNotRegisteredOrInActive')
             elif(UserDetails == None):
                 self.redirect('/UserSignIn?notification=EmailIdNotRegisteredOrInActive')
-            Latitude1 = UserDetails.Latitude
-            Longitude1 = UserDetails.Longitude
-            SignInStatus = "SignOut"
-            CartData = ndb.Key("CartDB",userEmail).get()
-            if(CartData != None):
-                for i in range(0,len(CartData.ProductID)):
-                    ProductData = ndb.Key("ProductsDB",CartData.ProductID[i]).get()
-                    ProductDetails.append(ProductData)
-            ProductDetailsLength = len(ProductDetails)
-            if(Button == "RemoveFromCart"):
-                ProductID = self.request.get("ProductID")
-                if(len(CartData.ProductID)>1):
+            elif(UserDetails != None and UserDetails.IsActive == 1):
+                Latitude1 = UserDetails.Latitude
+                Longitude1 = UserDetails.Longitude
+                SignInStatus = "SignOut"
+                CartData = ndb.Key("CartDB",userEmail).get()
+                if(CartData != None):
                     for i in range(0,len(CartData.ProductID)):
-                        if(CartData.ProductID[i] == ProductID):
-                            del CartData.ProductID[i]
-                            del CartData.Quantity[i]
-                            CartData.put()
-                            break
-                else:
-                    CartData.key.delete()
-                self.redirect('/ShoppingCart?userEmail='+UserDetails.user_Email+'&notification=ProductRemoved')
+                        ProductData = ndb.Key("ProductsDB",CartData.ProductID[i]).get()
+                        ProductDetails.append(ProductData)
+                ProductDetailsLength = len(ProductDetails)
+                if(Button == "RemoveFromCart"):
+                    ProductID = self.request.get("ProductID")
+                    if(len(CartData.ProductID)>1):
+                        for i in range(0,len(CartData.ProductID)):
+                            if(CartData.ProductID[i] == ProductID):
+                                del CartData.ProductID[i]
+                                del CartData.Quantity[i]
+                                CartData.put()
+                                break
+                    else:
+                        CartData.key.delete()
+                    self.redirect('/ShoppingCart?userEmail='+UserDetails.user_Email+'&notification=ProductRemoved')
         else:
             self.redirect('/UserSignIn')
 
@@ -66,6 +68,7 @@ class ShoppingCart(webapp2.RequestHandler):
                     Category.append(ProductsData[i].Category)
         Category.sort()
 
+        VendorProductsDetails = VendorProductsDB.query().fetch()
         PharmacyDetails = PharmacyDB.query().fetch()
         if(PharmacyDetails != None):
             for i in PharmacyDetails:
@@ -84,6 +87,7 @@ class ShoppingCart(webapp2.RequestHandler):
             'ProductDetails' : ProductDetails,
             'ProductDetailsLength' : ProductDetailsLength,
             'PharmacyDetails' : PharmacyDetails,
+            'VendorProductsDetails' : VendorProductsDetails,
             'CartData' : CartData,
             'Distance' : Distance,
             'notification' : notification,
@@ -106,6 +110,7 @@ class ShoppingCart(webapp2.RequestHandler):
             if(CartData != None):
                 Quantity = []
                 PharmacyID = []
+                OrderType = self.request.get('OrderType')
                 Valid = True
                 for i in range(0,len(CartData.ProductID)):
                     Key1 = "Quantity"+CartData.ProductID[i]
@@ -122,10 +127,15 @@ class ShoppingCart(webapp2.RequestHandler):
                         break
                     else:
                         PharmacyID.append(self.request.get(Key2))
-                if(Valid == True):
-                    CartData.Quantity = Quantity
-                    CartData.PharmacyID = PharmacyID
-                    CartData.put()
+                if(OrderType != None or OrderType != ""):
+                    if(Valid == True):
+                        CartData.Quantity = Quantity
+                        CartData.PharmacyID = PharmacyID
+                        CartData.OrderType = OrderType
+                        CartData.put()
+                        self.redirect('/ConfirmOrder?userEmail='+UserDetails.user_Email)
+                else:
+                    self.redirect('/ShoppingCart?userEmail='+UserDetails.user_Email+'&notification=InvalidOrderType')
         else:
             self.redirect('/UserSignIn')
